api/requirements.txt
flask
flask-cors
sqlalchemy
flask-sqlalchemy
pydantic
psycopg2-binary
pytest
requests
python-dotenv
openai
